OutputParsers/requirements.txt
langchain
langchain-core
langchain-openai
langchain-huggingface
python-dotenv
streamlit
pydantic
#pip install -r requirements.txt 